requirements.txt
et_xmlfile==2.0.0
openpyxl==3.1.5
psycopg2-binary==2.9.10
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.16.1
